Use native iOS websock if ios version 13
